--- src/main/java/nextstep/subway/domain/FarePolicy.java
@@ -1,5 +1,0 @@
-package nextstep.subway.domain;
-
-public interface FarePolicy {
-
-}
