+++ src/main/java/nextstep/subway/domain/fare/FareApplier.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare;
 
 public abstract class FareApplier {
