+++ src/main/java/nextstep/subway/domain/fare/FareApplierChain.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare;
 
 import lombok.AllArgsConstructor;
