+++ src/main/java/nextstep/subway/domain/fare/FarePolicy.java
@@ -1,0 +1,5 @@
+package nextstep.subway.domain.fare;
+
+public interface FarePolicy {
+
+}
